--- movie_rating_app/Models/AspNetRole.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace movie_rating_app.Models
-{
-    public partial class AspNetRole
-    {
-        public AspNetRole()
-        {
-            AspNetRoleClaims = new HashSet<AspNetRoleClaim>();
-            Users = new HashSet<AspNetUser>();
-        }
-
-        public string Id { get; set; } = null!;
-        public string? Name { get; set; }
-        public string? NormalizedName { get; set; }
-        public string? ConcurrencyStamp { get; set; }
-
-        public virtual ICollection<AspNetRoleClaim> AspNetRoleClaims { get; set; }
-
-        public virtual ICollection<AspNetUser> Users { get; set; }
-    }
-}
