--- movie_rating_app/Models/AspNetRoleClaim.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace movie_rating_app.Models
-{
-    public partial class AspNetRoleClaim
-    {
-        public int Id { get; set; }
-        public string RoleId { get; set; } = null!;
-        public string? ClaimType { get; set; }
-        public string? ClaimValue { get; set; }
-
-        public virtual AspNetRole Role { get; set; } = null!;
-    }
-}
