--- movie_rating_app/Models/AspNetUser.cs
@@ -1,46 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace movie_rating_app.Models
-{
-    public partial class AspNetUser
-    {
-        public AspNetUser()
-        {
-            AspNetUserClaims = new HashSet<AspNetUserClaim>();
-            AspNetUserLogins = new HashSet<AspNetUserLogin>();
-            AspNetUserTokens = new HashSet<AspNetUserToken>();
-            Favourites = new HashSet<Favourite>();
-            Reviews = new HashSet<Review>();
-            Roles = new HashSet<AspNetRole>();
-        }
-
-        public string Id { get; set; } = null!;
-        public string? UserName { get; set; }
-        public string? NormalizedUserName { get; set; }
-        public string? Email { get; set; }
-        public string? NormalizedEmail { get; set; }
-        public bool EmailConfirmed { get; set; }
-        public string? PasswordHash { get; set; }
-        public string? SecurityStamp { get; set; }
-        public string? ConcurrencyStamp { get; set; }
-        public string? PhoneNumber { get; set; }
-        public bool PhoneNumberConfirmed { get; set; }
-        public bool TwoFactorEnabled { get; set; }
-        public DateTimeOffset? LockoutEnd { get; set; }
-        public bool LockoutEnabled { get; set; }
-        public int AccessFailedCount { get; set; }
-        public string? FirstName { get; set; }
-        public string? LastName { get; set; }
-        public int? Nationality { get; set; }
-
-        public virtual Nationality? NationalityNavigation { get; set; }
-        public virtual ICollection<AspNetUserClaim> AspNetUserClaims { get; set; }
-        public virtual ICollection<AspNetUserLogin> AspNetUserLogins { get; set; }
-        public virtual ICollection<AspNetUserToken> AspNetUserTokens { get; set; }
-        public virtual ICollection<Favourite> Favourites { get; set; }
-        public virtual ICollection<Review> Reviews { get; set; }
-
-        public virtual ICollection<AspNetRole> Roles { get; set; }
-    }
-}
