--- movie_rating_app/Models/AspNetUserClaim.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace movie_rating_app.Models
-{
-    public partial class AspNetUserClaim
-    {
-        public int Id { get; set; }
-        public string UserId { get; set; } = null!;
-        public string? ClaimType { get; set; }
-        public string? ClaimValue { get; set; }
-
-        public virtual AspNetUser User { get; set; } = null!;
-    }
-}
