--- movie_rating_app/Models/AspNetUserLogin.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace movie_rating_app.Models
-{
-    public partial class AspNetUserLogin
-    {
-        public string LoginProvider { get; set; } = null!;
-        public string ProviderKey { get; set; } = null!;
-        public string? ProviderDisplayName { get; set; }
-        public string UserId { get; set; } = null!;
-
-        public virtual AspNetUser User { get; set; } = null!;
-    }
-}
