--- movie_rating_app/Models/AspNetUserToken.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace movie_rating_app.Models
-{
-    public partial class AspNetUserToken
-    {
-        public string UserId { get; set; } = null!;
-        public string LoginProvider { get; set; } = null!;
-        public string Name { get; set; } = null!;
-        public string? Value { get; set; }
-
-        public virtual AspNetUser User { get; set; } = null!;
-    }
-}
