+++ movie_rating_app/Models/Nationality.cs
@@ -8,7 +8,7 @@ namespace movie_rating_app.Models
         public Nationality()
         {
             Actors = new HashSet<Actor>();
-            AspNetUsers = new HashSet<AspNetUser>();
+      //      AspNetUsers = new HashSet<AspNetUser>();
             Creators = new HashSet<Creator>();
         }
 
@@ -16,7 +16,7 @@ namespace movie_rating_app.Models
         public int Id { get; set; }
 
         public virtual ICollection<Actor> Actors { get; set; }
-        public virtual ICollection<AspNetUser> AspNetUsers { get; set; }
+   //     public virtual ICollection<AspNetUser> AspNetUsers { get; set; }
         public virtual ICollection<Creator> Creators { get; set; }
     }
 }
