+++ movie_rating_app/appsettings.json
@@ -1,6 +1,6 @@
 {
   "ConnectionStrings": {
-    "DefaultConnection": "server=DESKTOP-AMGTN3U;Database=aspnet-movie_rating_app;Trusted_Connection=True;MultipleActiveResultSets=true"
+    "DefaultConnection": "server=DESKTOP-AMGTN3U;Database=aspnet-movie_rating_app_2;Trusted_Connection=True;MultipleActiveResultSets=true"
   },
   "Logging": {
     "LogLevel": {
